+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Flyout: Using vanilla event listeners on trigger, you can pass additional custom handlers yourself
 - View: Responsive textAlign support
 - Toast: Warning color support
+- FileUpload: Added `height` support